requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
pgvector==0.3.6
supabase==2.25.1
python-multipart==0.0.20
httpx==0.27.2
google-generativeai
openai==0.8.3
